add loop back device and btrfs file system creation support
To test btrfs related functionality we need a test harness that should
be as isolated as possible from the rest of the system. To achieve this
level of isolation, using a loop back device appears to be a good
option.
This change introduces the test harness for future btrfs related
functionality. It provides objects for creation of loop back devices,
formating them with btrfs, and mounting them properly. To not run into
clashes with other components, temporary files and directories are used
where required.