introduce variable aliasing functionality
When writing Python code sometimes long variable names are unavoidable:
member variables should be reasonably expressive, for instance, and
accesses are always prefixed with a 'self.'.
This change introduces an alias function that can be used to provide
an additional name for a variable within a 'with' block. Example usage
could look like:
  self._incredibly_long_variable_name = Object(...)
  with alias(self._incredibly_long_variable_name) as var:
    var.doSth()
    var.doSthElse()