refactor btrfs functionality to only return lists of commands
Certain programs run best in a pipelined fashion. A lot of UNIX programs
are connected over their stdout/stdin file descriptors and might form a
chain of commands. The upcoming send and receive functionality will
follow this paradigm for it appears to be the most efficient way of
achieving the wished for result. The current approach of directly
executing the desired btrfs command does not (easily) allow for this
kind of chaining.
To that end, this change modifies the btrfs module to merely return
lists containing commands and their arguments instead of directly
executing them.

Note: A different approach is to handle command pipelining entirely in
      Python, i.e., reading one process' output into a Python byte array
      and passing it to the next program.
      The problem there is that all the Python variable handling comes
      at a cost (allocations, garbage collection, and more) and since we
      might send large amounts of data (an entire snapshot), we want
      this process to have as low overhead as possible. By using the
      native pipelining interface (i.e., at the system call level), we
      do not add any additional abstractions over shell-style pipelines.
      Furthermore, the additional flexibility provided by being able to
      handle the data directly in Python is not required for our use
      cases.