introduce pipelining functionality
In order to send and receive snapshots in the most flexible way, we need
to be able to create a pipeline of commands. That is, in the simplest
form the btrfs send and receive commands work like this:
  btrfs send <snapshot> | btrfs receive <destination>

This change introduces the necessary pipeline execution functionality to
be able to chain arbitrary commands by connecting their stdin and stdout
file descriptors.

import subrepo ./:execute at 79f080ba615aa0da0089acde989ce563a0d37199