introduce serialization and deserialization infrastructure
One of the desired outcomes of this application is to be able to backup
btrfs subvolumes locally *or* remotely. To achieve this goal, the btrfs
send and receive functionality has to be harnessed which allows for
creating file-system supported low-overhead incremental backups as well
as for sending and receiving them.
This change introduces the required functionality that can be used for
converting a snapshot into a byte stream, transferring it somewhere, and
possibly recreating a btrfs subvolume from this stream. Note that the
latter step is optional and the backing store could also be an arbitrary
file in which the data can be restored (in which case any file system
could be used -- not necessarily btrfs as is the case when the subvolume
is truly recreated).