introduce sync functionality
Various sources indicate that snapshots have to be physically persisted
to disk before they can be sent (i.e., serialized), otherwise "stale NFS
file handle" errors might occur.
Apparently, this is a known issue so it is supposed to be fixed at some
point (perhaps it is already and the sources are just outdated but I was
unable to find a mentioning of it being no longer an issue). However, it
is better not to take chances when it comes to backups and so this
change introduces the functionality to perform a btrfs sync to disk.