add command support for serializing incremental snapshots
The first step towards incremental synchronization of repositories is to
have the proper command support hooked up, i.e., to have the wrappers
for the btrfs program in place.
This change adjusts the serialize() command generation function to
accept an additional (and optional) parameter that specifies the parent
to use, i.e., the snapshot to take as a base when sending only
incremental changes.
It is worth pointing out that snapshots in btrfs are, by definition,
incremental (internally), because changes are made in a copy-on-write
fashion. As such, adjusting the serialize() function to be able to
create a command that relates a given snapshot to a parent one is all
that is required to send incremental changes on a low level.