allow for multiple parent snapshots to be used
The most recent change added support for transfering incremental changes
between two snapshots only in order to sync up two repositories. The
approach we used had two minor flaws:
1) We did only specify one parent to be passed to the btrfs command. Although
   one parent, as long as it is the most recent one, is enough to create a
   small incremental snapshot, better sharing on the destination repository's
   btrfs file system can occur if multiple clone sources are specified. Clone
   sources in these terms can be all previously created snapshots.
2) If a most recent snapshot could be found on the source side, there is no
   guarantee that it is available on the destination side as well (although
   this should be the case if the user did not manually delete created
   snapshots on the destination repository). In case this most recent snapshot
   does not exist on the destination, the entire subvolume has to be
   serialized and transferred. This is suboptimal in the sense that there
   might be other, less recent snapshots available on both sides that could be
   very well used to transfer only an incremental change.
This change addresses both issues. It modifies the serialize() command to
accept a list or set (or any iterable sequence) of parents. If specified, they
are all passed to the btrfs command by means of the clone source (-c) option.
Furthermore, the repository from now on uses all snapshots that are available
in both repositories and passes them in as the parents. This way, we do not
artificially limit btrfs in terms of what data can be shared.