provide a main function and a btrfs-backup script
Until now btrfs-backup could not be considered a true program. There was
no command-line interface available and so interaction could only be
done directly using Python.
This change introduces the functionality to accept and interpret user
supplied commands as well as to act on them. It adds a btrfs-backup
script that is meant to be the starting point of execution, i.e., that
is the part to be invoked directly by the user.
As of now the user has to provide the program with a source repository,
a destination repository, and a number of subvolumes (using one or
more instances of the -s/--subvolume parameter).