add packaging functionality
This change adds support for packaging and distributing btrfs-backup using
the setuptools package. Providing the means for packaging in terms of
Python infrastructure compared to at the operating system distribution
level has the advantages that
  a) we are independent of distributions and can even use the native
     Python way of distributing things and
  b) it simplifies creation of distribution specific packages because
     most distributions or package managers provide means to interface
     with setuptools packaged Python modules in a generic fashion.

More subtle additions of this change are the addition of __init__.py
files to the major modules (otherwise we see warnings popping up) and
the introduction of loader scripts that can be installed and invoked
directly.