add support for prefixing the temporary test directory
We use temporary files and directories as the means for backing and
mounting loop back devices. This approach avoids clashes with files with
similar names over the case when fixed names would be used and also
allows for running tests in parallel (although this is currently not
done).
However, until now we were unable to honor things such as sandbox
environments in which the tests are run and from which they should not
break out. This change wraps the existing temporary file and directory
creation functions and checks for an environment variable, TEST_TMP_DIR,
that specifies a directory into which to create the temporary objects.
Using this approach, simple isolating sandbox environments which are
located somewhere in the file system tree will not be violated by our
tests anymore.