ensure uniqueness of snapshot names with respect to their paths
Until now we had the problem that snapshot names for subvolumes are not
unique with respect to paths in non-canonical forms. That is, symbolic
links, relative paths, and other constructs used in a subvolume's path
would result in different snapshot names being used which is a source
of unconvenience in many respects.
This change fixes the issue by using canonical path names to create the
snapshot's name. In addition, with this change the leading path
separator in absolute paths which resulted in a needless underscore in
the derived snapshot name is removed.