add tests to verify non-existent directories are handled correctly
When specifying invalid input, for instance, by using invalid paths to
subvolumes or repositories, the program should fail in a way that makes
apparent what the issue is but also not only detect this problem in the
middle of an important operation and leave over data that is in a
potentially inconsistent state.
This change adds tests cases that check for these properties under some
cases. Note that the decision was made to not explicitly check all input
parameter for validity, e.g., by looping over all passed in subvolumes
and checking if they really exist before synchronizing them. The reason
is that this would pollute the code more than what we would actually
gain as a benefit. Also, such checks are always just best effort, even
if performed early on start up the directories/subvolumes could be
removed between this initial check and the actual process of taking a
snapshot.