fix reported changed files by diff
Recent testing discovered that when creating a list of changed files
since the last snapshot, the files that changed as part of the snapshot
are included as well.
The reason for this behavior is a supposed bug in btrfs's generation
handling in conjunction with the find-new command (it might also just be
a strange or unintuitive behavior in conjunction with insufficient
documentation).
To work around this issue, we increment the generation ID reported for a
snapshot by one before passing it to the find-new commmand. Since in
general the command accepts arbitrary generations (i.e., non-existent
ones as well) without reporting an error, we should be safe. A test case
was added for this very case. The existing tests cover any potential
false negatives, i.e., cases where increasing the generation would
result in no changes being detected although they exist.