introduce restore functionality
Until now it was possible only to backup data. The process of
restoration had to be done manually.
This change introduces support for automated restoration of snapshots.
Note that for now we do not recreate the original subvolume from which
the snapshot was taken -- such an approach is subject to a future
change. Restoration can be triggered by supplying the -r/--restore
option to the program. In such a case the script searches for the most
recent snapshot in the given source repository and copies it to the
destination (if it does not already exist).