add full subvolume restoration support
Until now it was only possible to restore the snapshots from a backup.
The assumption was that the user would be able to restore the subvolume
from the latest snapshot himself/herself. Obviously, this process can be
automated as well.
To that end, this change modifies the restoration process to recreate
a writable subvolume out of the latest snapshot that was restored from
the backup. In order to achieve the previous behavior of only restoring
the snapshots from the backup, the newly introduced --snapshots-only
option can be supplied.