add support for automated deletion snapshot of old snapshots
Up to this point the btrfs-backup was only able to create and deploy
snapshots. Over time, the snapshots accumulate and manual deletion
should be performed to get rid of some of the unused ones in order to
save space and to preserve overview.
This change establishes a way for automated removal of old snapshots. By
means of the newly introduced --keep-for option a duration can be
specified that defines for how long snapshots are kept in the local as
well as the remote repository. If a snapshot is older than the current
time minus the given duration, the snapshot will be deleted when the
next backup is performed. Note that as a safety measure the most recent
snapshot will never be deleted.
A duration as accepted by the new option is a number and a unit of time.
Supported duration units are seconds (S), minutes (M), hours (H), days
(d), weeks (w), months (m), and years (y).