fix problem where restoration succeeded although the subvolume existed
This change fixes a problem where restoration of a subvolume succeeded
although the subvolume did actually exist. The problem was that we
relied on the btrfs command to detect when a snapshot/subvolume existed
in which case it should report an error. However, if there exists a
directory with the name of the subvolume to restore, restoration would
just happen *inside* this directory.
With this change a special case is added that checks for existence of a
directory with the name of the subvolume to restore. If one exists, we
explicitly raise an exception.
Note that we explicitly special case directories only. If a file happens
to be where the subvolume is to be remanifested, the btrfs program will
detect this issue.