introduce --no-read-stderr option
When executing programs I experienced a subtle problem that is worth
pointing out. In my particular case I used the previously introduced
remote command execution support (by means of the --remote-cmd option)
to create a backup of local data on a remote host. Specifically, I used
SSH to connect to the host of interest. In my setup, SSH is configured
to run in so called ControlMaster mode [1], which means that a master
connection is established upon first contact with the host and
subsequent connections to this host can be multiplexed over the master
connection.
Now, what SSH does when no master connection exists is to establish a
new connection and fork into the background (I am assuming the
ControlPersists option is set as well). The initially created process
uses this master connection for the execution of the supplied command
and when it is done the master connection can persist as the separate
process that forked off.
When backing up the data to a different host using SSH a delay of 1
minute was incurred for every executed remote command if, and only if
the backup was the process to create the master connection (i.e., no
master connection was present).
Investigation has shown that the reason for this behavior is that by
default we read all available data from stderr in order to be able to
create a proper error message (by raising an exception) in case the
command failed. The interesting part is now that SSH, once it forks into
the background, apparently closes stdout but leaves stderr open. Since
both SSH processes are connected to btrfs-backup, the pipe used for
reading data from stderr stays open as long as one of the child
processes is alive, that is, until the process owning the master
connection dies. For me, this is exactly the 1 minute delay I witnessed
because of a ControlPersist value of exactly that.
Arguably, there is no general solution to this problem. We want to
include error output in the normal case to get meaningful error messages
in case something goes wrong. We certainly cannot modify SSH and neither
can a configuration change be expected for users that want to use the
remote command functionality.
To that end, this change introduces a new option, --no-read-stderr that
can be used to disable the collection of data from stderr which means in
case something goes wrong only the error code is available for
diagnosis.
As an aside: the simplest way to work around this problem in the first
place is to just establish the master connection before running
btrfs-backup, or simply not using ControlMaster mode at all.

[1] man 5 ssh_config