introduce helper functionality for --reverse option
With an upcoming change we need to interact with the source and
destination repository as well as the send and receive filter arguments
as parsed by the ArgumentParser functionality at two places. Because we
have the --reverse option, handling gets complicated since the reversing
code needs to be duplicated.
This change adjusts the way the --reverse option achieves its goal.
Instead of manually reversing the options after argument parsing is done
we introduce an undocumented helper option (--reverse-hidden-helper).
Compared to --reverse it accepts an (ignored) argument. This argument
will be parsed by the argument parser and this is where we hook in. When
the parsing happens we transparently reorder the arguments of the
namespace, a reference to which we got passed in earlier. In order for
this mechanism to work, the newly introduced hidden helper function
needs to be inserted as the last option of the argument vector because
the ArgumentParser parses arguments in the order they appear in the
vector. To that end, we furthermore provide functionality capable of
performing this insertion at the right position (taking into account
possible '--' strings separating options from arguments).