introduce --snapshot-ext option
One of this program's goals is to support not only native btrfs volumes
for storing snapshots but also arbitrary file systems. In such a
scenario, a snapshot cannot be remanifested but instead the serialized
stream has to be stored in a file.
To support this architecture, we need a way to tell the program that
storing the data in a file is desired as opposed to actually
deserializing the stream produced by the btrfs program and recreating
the snapshot/subvolume.
This change introduces the --snapshot-ext option (short for snapshot
extension). The idea is that once we store snapshots in a file we want
to provide a file extension that identifies the data in the file. The
reason for this behavior is that once we apply various filters (by means
of the --send-filter or --receive-filter option), data can get processed
in arbitrary ways: compression and encryption certainly are upon the
transformations that come to mind. The option is not yet properly hooked
up, i.e., it does nothing except checking for various prerequisites to
work properly. One of the constraints it enforces is that the last
receive filter (in case of a backup) or the first send filter (in case
of a restore) contain a '{file}' string which is subsequently replaced
by the snapshot file's name.