allow Repository to work on relative paths
A previous change adjusted the Repository class to work with absolute
paths, even if a relative one was passed in. With the introduction of
remote directories this approach is no longer feasible: in general we
want to avoid executing (and thus, relying, on) anything except what is
truly necessary (think: the btrfs suite of programs) on the remote end
of the connection. This includes utilities such as realpath and whatnot
to canonicalize paths.
This change allows for relative paths to be accepted on various
repository based operations.