add support for supplying file descriptors to execution functions
Until now the execute() and pipeline() functions supported the following
scenarios: for input, data could be supplied from a Python context (in
the form of a bytes-like object) or, if no data was given, it was
implicitly read from a null device. For output, it could be stated
whether or not data from the started process (or processes) should be
read.
There is one major (artificial) limitation that became apparent: it is
not possible to supply a file descriptor to use as stdin, stdout, or
stderr.
This change allows for file descriptors to be passed in to the execution
functions. Furthermore, the non-orthogonality between the 'data_in' and
'read_*' keyword parameters to the two functions (both in terms of the
names and the accepted parameter types) was removed. All keyword
parameters (which are now named stdin, stdout, and stderr) now can
either be None, a file descriptor, or a bytes-like object. When
supplying data to the stdout or stderr parameters, this data will end up
initializing the buffer into which all other data from the pipeline is
read, effectively prepending it to the actual output.  Arguably, there
are not many use cases where this behavior comes in handy but it
simplified the code quite a bit.

import subrepo ./:execute at 055ea928b34dbac9ef8b2f7580ae4711dbfb26aa