Add functionality for creating loop back devices
In order to test various aspects of our program we are going to need to
work on an actual btrfs file system. To make that possible, this change
introduces some easy to use wrappers around the losetup command that
allow for the management of loop back devices, which will form the basis
for creating temporary file systems.