Add support for crafting necessary btrfs-progs commands
btrfs functionality such as the sending and receiving of subvolumes does
not come in a library, but rather is provided by btrfs-progs and the
btrfs command specifically. For us that means that instead of having a
proper API we have to craft commands that instruct the btrfs program to
do what we want it to.
With this change we introduce the necessary functions with an internal
API that we can program against.