Introduce concept of a repository
A lot of the logic for dealing with btrfs snapshots will revolve the
concept of a repository, which is basically a directory on a btrfs file
system that houses snapshots. It can be the source or the destination of
a subvolume send operation.
This change introduces the Repo type that represents such a repository.
Currently, all it is able to do is to find the root of the underlying
btrfs file system, which will be required for various operations.