Add support for parsing and creating snapshot names
Whenever we create a snapshot of a subvolume, we encode a bunch of
information into its name. This information will later help us identify,
among other things, which subvolume it was created for.
With this change we introduce the Snapshot type along with supporting
infrastructure. This type encapsulates all the information contained in
such a name, and it can be used to parse or emit one.