Add support for creating and listing snapshots
This change introduces the means to create a snapshot of a subvolume
inside a repository as well as to list all snapshots inside of it.
Listing of snapshots happens in parsed form and we ignore all snapshots
(and btrfs subvolumes in general) whose naming pattern we do not
recognize -- these are effectively not managed by our repository
abstraction.