Introduce command line argument handling infrastructure
The program will be command line driven, with top-level sub-commands
indicating the operation to be performed (e.g., backup or restore). With
this change we start laying the foundation for doing so. Specifically,
we introduce the means of parsing program arguments via clap and
introduce a first sub-command for creating a snapshot of a subvolume.