Add support for printing executed commands
For debugging purposes or just to get a better understanding of what the
program is doing, it can be useful to get a trace of all commands being
executed.
This change introduces the infrastructure for doing so via the newly
introduced --trace option. On the implementation side, we add a global
flag to the btrfs module, which is the only module executing commands
(aside from test code). If need be, we could control this setting on a
per-instance basis, but given the intent this seems overkill and
needlessly added complexity.