Canonicalize subvolume paths correctly
When users backs up or restore a subvolume, they would reasonably
expect that different paths to the same volume (e.g., via relative path
specifications or through different symbolic links) are resolved
in such a way that backup is the most efficient (no unnecessary
snapshots are created) and that restoration works correctly.
Currently that is not the case, because we treat user provided paths to
the subvolume verbatim.
With this change we fix this issue, by canonicalizing file system paths
properly on the relevant code paths.