Explicitly pass in program arguments to run() function
So far our run() function relied on a bunch of implicit state in the
form of global program arguments that were taken from the environment.
That does not lend itself to proper testing.
In an attempt to add a set of end-to-end tests with upcoming patches,
with this change we adjust the function to have the program arguments
passed in instead.