Add end-to-end tests for existing functionality
We have reasonable code coverage for the "core" btrfs functionality.
However, a fair amount of logic is concerned with merely acting on
program arguments and the plumbing for that is effectively untested.
That makes it hard to put faith into the argument handling.
With this change we add a bunch of end-to-end tests that effectively
mostly make sure that the existing sub-commands as well as various
option combinations to them result in plausible state changes.

We introduce these tests as integration-style ones, working on the
crate's public API surface. As such we need to export a few bits of
functionality publicly. The main reason for going with integration style
tests (as opposed to unit tests) is that just conceptually we are
testing at the highest level possible, using the very command line
interface that the crate provides. As such, testing against the public
API surface seems the most sensible, although that distinction is mostly
irrelevant given that the crate at hand is an application and not a
library in any stricter sense.