Make encoded subvolume paths first class citizens
Our Snapshot type represents a snapshot for a subvolume. The path to
said subvolume is encoded in its name. So far we have always decoded the
subvolume's path and stored it inside the corresponding Snapshot object.
This decoding step requires the encoding to be loss-less.
Currently, however, because we do not escape underscores in path
components, it is not. For example, both /snapshots/xxx_yyy and
/snapshots/xxx/yyy would result in the same encoded name. This can cause
subtle issues where subvolumes and snapshots do not match up although
they should.
To fix this problem, this change introduces a dedicated Subvol type
which represents the encoded subvolume path and never decodes it back.
In addition to that, the new type more tightly controls what operations
are possible and currently we do not expose any mutating calls
(something very well possible on the previously used PathBuf). That is
somewhat important, because we expose the corresponding Snapshot::subvol
member publicly.