Remove os and hw components from snapshot names
The "os" and "hw" components, which we include in snapshot names, are
not believed to actually provide essential information. It is utterly
unlikely that these attributes really add any differentiating
information on top of what the already included hostname provides. That
is to say, we do not believe that it is good practice to name hosts the
same but have them only differ in operating system or hardware type.
That could conceivably happen in a dual-boot setup, in which case "os"
could be variable but "hw" would be constant, but BTRFS is a Linux-only
technology and so that would different distribution usage and we would
not be able to detect that currently either.
To that end, this change removes the "os" and "hw" components from
snapshot names altogether, making them shorter and easier on the eyes.