Introduce ENCODED_PATH_SEPARATOR constant
Currently our encoding and decoding logic for snapshot names is somewhat
convoluted by the fact that we are hard coding the various component
separators in a few places.
As a first step towards improving the situation, this change introduces
the ENCODED_PATH_SEPARATOR constant, which represents the character we
use to encode path component separators in a file system path.