Create date and time formats more programmatically
The usage of the format_description macro for creating date and time
formats is arguably problematic: it prevents us from actually
referencing the separator constants that we introduced earlier while
also being macro magic running at compile time.
With this change we replace the usage of this macro with compile time
crafted arrays of FormatItem components. Doing so is more flexible and
to the point than using the format_description macro and it allows us to
more easily use the same formats for parsing and formatting.