Swap snapshot name separators
So far our generated snapshot names had used a dash as the main
component separator while intra-component separation was (mostly) done
with an underscore.
This change flips usage around so that all top level components are now
separated by an underscore. Doing so makes everything more uniform in
the presence of dates that we want to keep separating by dashes.