Introduce logic for escaping and unescaping strings
Our current snapshot naming logic still has one problem: we can easily
get parsing errors if subvolume names contain characters we use for
separating the snapshot name components. To fix this issue, we are going
to need to escape those characters properly.
This change introduces the functionality necessary for escaping and
unescaping, but does not yet hook it up.