Introduce non-strict canonicalization functionality
Rust's std::fs::canonicalize() function is rather strict when it comes
to accepting input paths and reporting errors. Specifically, if a path
does not exist completely, but part of it does, it will report a
file-not-found error. Such behavior is not great in some scenarios where
what is really desired is a best-effort canonicalization of all the
components that exist on the file system, followed by the non-existent
remainder.
This change introduces the canonicalize_non_strict() function that
ignores non-existent parts of a path while canonicalizing everything
else. The function is used to simplify parts of the code for restoring
of snapshots, but it will become essential with upcoming changes that
introduce the means for purging old snapshots.