Add bindings for making a subvolume read-only
With upcoming changes we would like to create fake snapshots for testing
purposes. Given that a snapshot in our program's sense is really only a
read-only volume with a special name, the only piece missing for the
creation of such snapshots is the means for making a subvolume
read-only.
This patch introduces the necessary plumbing for changing that state of
a subvolume.