Add support for deleting snapshots from the destination repository
So far we have only supported "purging" of snapshots on the source file
system. However, it is entirely reasonable to also want to keep the
number of backups limited.
This change introduces the means for managing snapshots in the
destination repository as well, by adding the --destination option to
the purge sub-command.