Introduce GitHub Action based CI
This change adds a configuration file for a GitHub Actions based CI
pipeline to the repository. This file (directly or indirectly) controls
the environment in which to build and test, and is also used for some
additional linting.
With the pipeline in place the change also adds a badge indicating the
status of the build on the project's front page.