Fix --version/--help options to not error out
The clap crate has this counter-intuitive behavior of the
Parser::try_parse_from() function to return an error when the user
requested the display of a help text or version string (among other
things). Because we missed that, we unintentionally reported errors when
the user provided the --help or version option.
This change fixes the issue by exiting successfully instead.