Use proper destructuring in Snapshot's Display impl
The Display impl for the Snapshot type is supposed to capture all
members of the type. Currently, it just accesses members of "self" by
name, but if the type were to be extended with additional members, that
code could silently do the wrong thing.
With this change we destructure the type properly, making sure that all
its members are evaluated at compile time.