Add tag member to Snapshot type
This change introduces a new member to the Snapshot type: tag. The tag
is a new mandatory component of each snapshot name (though it may be
empty). Once fully wired up in the future, it will allow snapshots of
the same subvolume but for different backup destinations to be
distinguished, which can help ensure that snapshots only present on a
certain backup medium are not accidentally deleted from the source,
necessitating a full backup next time around.