Allow for tagging of snapshots
Depending on the users workflow, it may make sense to backup a subvolume
to different destinations at different intervals. E.g., a subvolume
could be backed up frequently as part of an "important" backup cycle and
irregularly as part of a full system backup. If snapshots and backups
are managed via the purge command, which exclusively considers
timestamps when making purging decisions, the snapshots for the
irregular full system backup are likely to have been removed. As such,
the incremental nature of backups would be lost. That is a very
unfortunate behavior and it is close to impossible for users to work
around that in a sensible way.
With this change we introduce the means for tagging snapshots to give
users the necessary tools to deal with such scenarios. A tag in our
sense is any free-form string and it included as a suffix in snapshot
names when creating a snapshot or pulling a backup. When purging, we now
only remove snapshots with the same tag, meaning that if users tag
snapshots taken for different backup destinations differently, they can
prevent "old" snapshots with a different tag from getting purged.