Introduce FileOps trait abstracting away file system operations
With upcoming changes we would eventually catch up to the original
btrfs-backup's functionality when it comes to remote backups. For that
to happen, we need to be able to perform a certain set of file system
operations, for which we currently use local function and system calls,
on a remote system. The way this will happen is via command executions
that fulfill a similar function.
To abstract away the differences between the calls necessary on the
local system and those required for a remote one, this change introduces
the FileOps trait. This trait captures all operations currently required
for the core btrfs backup functionality. With subsequent changes an
instance of the LocalOps type, which implement this type for local file
operations, will be used throughout the code base.