Add support for prefixing btrfs commands
We eventually want to support backing up btrfs subvolumes to remote
systems, e.g., over ssh. The way we envision that to happen is by
allowing users to provide a command that establishes such a remote
connection and which accepts arguments that will be executed on it. We
could then execute our btrfs commands on the remote system in this
manner.
This change introduces the first bit of plumbing to get to that point.
Specifically, it adds logic to the Btrfs type to honor such a remote
command and prefix all btrfs invocations with it.