Correctly assemble btrfs commands honoring "prefix"
With this change we adjust the Btrfs type to correctly prefix all
executed commands with the (optional) remote command provided at
construction time (though the plumbing for providing this prefix does
not yet exist). The one current exception is the send_recv() method,
which creates a pipeline of commands that requires additional API
changes. It will be adjusted subsequently.