Introduce builder infrastructure for Repo type
In the future we would like to make more aspects of a Repo object
configurable.
To set the foundation for doing so, this change introduces the
RepoBuilder type, which can be used to construct a Repo object. Right
now all that can be built is a "default" version of the object, but with
upcoming changes additional parts will be configurable via this builder
type.