Introduce RemoteOps type for file operations on a remote system
As motivated earlier, with upcoming changes we need the ability to
perform certain file system operations on a remote system, as opposed to
locally. We introduced the FileOps trait in order to abstract away those
details at callsites and the LocalOps type implementing said trait to
provide local file operations.
With this change we introduce the counterpart for providing remote
file operations: RemoteOps. In a nutshell, we map file operations to
executable commands and invoke them on the remote.