Move subvolume canonicalization into command handling logic
Our core "repo" functionality currently canonicalizes subvolume paths
freely. That is not wrong at this point, as our assumption is that we
are working with canonical paths anyway. However, once we fully support
working on remote systems, it will make a difference whether we
canonicalize locally or remotely.
Subvolume paths, specifically, should always be canonicalized on the
local system, as that acts as the reference. A remote repository should
treat those paths in a verbatim manner.
To make that more explicit, this change moves the canonicalization logic
from the repo module into the command handling logic inside lib.rs.