Introduce MountBuilder type
This change makes our mounting infrastructure more generally usable by
introducing the MountBuilder type. This type can be used to easily
configure certain aspects of the mount. In the process we also make the
Mount type work with an existing directory, as opposed to
unconditionally creating a temporary one and mounting the device there.