Add tests for remote command functionality
This change adds tests for the remote command functionality, recently
introduced in the form of the --remote-command option added to various
of the program's sub-commands. We do not really have a remote system and
absolutely do not want to perform a live backup over SSH with
credentials injected somehow. As the next best thing, we create a chroot
with all necessary dependencies and make the chroot step the remote
command. This way, for the purposes of our tests, we get full isolation
from the actual system under test's file system.