Warn on unwrap in product code
As a general rule of thumb, we do not want to have any unwraps in
product code to avoid potential panics at runtime. However, sometimes
these are basically unavoidable.
With this change we warn on all unwraps outside of test code. Because we
have a few locations where we intentionally unwrap, we convert these to
expect calls to both circumvent the warning but also to provide at least
a bit more information should things really go sideways. That may come
in handy given that we abort on panic, meaning a backtrace is unlikely
to be available.