Fix handling of root directory subvolumes
As it turns out our handling of / as the subvolume is subtly broken.
Specifically, we fail to parse the corresponding snapshot file names as
valid snapshots and then silently end up ignoring them altogether. The
root cause of this problem is that we encode the root path '/' as an
empty subvolume string. When used in a snapshot name, because the
subvolume is effectively omitted, we end up with two component
separators right next to each other. This in turn is how we encode an
escaped component separator, which then means we fail to parse a
snapshot name correctly.
To fix this problem, this change adjusts the subvolume encoding to emit
a single intra-component separator when the path is '/'. As a result, we
change the way we encode the root directory altogether. That, however,
only affects newly created subvolume, which means that we rely on a
manual rename of existing snapshots of the root subvolume from
<host>__<date>[...] to <host>_-_<date>[...] so that they are recognized
correct in the future.