Don't create feature for 'tempfile' dependency
We use the 'test' feature to enable test related functionality. One of
its dependencies is the tempfile dependency. Currently, the way this
dependencies is specified causes an implicit crate feature to be created
as well. This has been recognized as a bad design decision by the Cargo
team. To that end, the next edition of Rust will change this behavior
and stop the creation of implicit features. With this change we opt into
this new behavior via the 'dep:' syntax, which in turn has been stable
for a while.