Attach statically build binary to release
Since forever we have wanted to attach statically linked binaries of
the program to the corresponding release. However, because GitHub
Actions can't seem to get their Matrix workflows interacting with
anything, really, that seemed impossible to do.
This change proposes a workaround that makes each job of the matrix
attach the binary to the release itself, provided that the associated
input argument it set.