Build static linked binary for armv7-unknown-linux-musleabihf target
On top of the x86_64-unknown-linux-musl one, we would also like to have
a statically linked binary for the armv7-unknown-linux-musleabihf
target. Adjust the build workflow to produce one. In order to make that
happen, switch to using cross compilation in a potentially emulated
environment using the taiki-e/setup-cross-toolchain-action GitHub
Action.