Inquire uname information only once
Currently we inquire the uname information from the system whenever we
create a SnapshotBase object. That is wasteful in the sense that this
operation is comparatively expensive, but it could also be a correctness
issue if the hostname changes from one point to the next, which isn't
really a possibility that we consider anywhere.
Circumvent these potential problems by gathering this information only
once and then reusing it. Contrary to the UTC offset, we don't panic
immediately on error but defer that to runtime once the host name is
actually required, to be as resilient as possible to potential errors.