Enable 'test' for devel builds
It is rather annoying that on a regular 'cargo test' run, the
integration tests in tests/end-to-end.rs are not actually run, because
they require the 'test' feature to be enabled, but it is not enabled by
default.
Make it so that on various devel builds, said feature is enabled, to
eliminate this annoyance.