Extend chroot copy logic to handle shell scripts
Some system commands may not be actual ELF files, but could just be
shell script. Coreutils, for example, supports a configuration in which
individual commands are just shell scripts of the form:

  > #!/usr/bin/coreutils --coreutils-prog-shebang=XXX

Add support for handling such files to the logic resolving dependencies
and copying them into the created chroot.